requirements.txt
requests==2.25.1
cloudinary==1.24.0
Pillow==8.1.2
numpy==1.20.1
opencv-python==4.5.1.48
dlib==19.21.1
python-dotenv==0.15.0
reportlab==3.5.59